@@ -1,8 +1,8 @@
 package tests.jvisick.backroom;
 
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.google.android.material.snackbar.Snackbar;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
@@ -36,7 +36,8 @@ public class RapidActivity extends AppCompatActivity {
         setContentView(R.layout.activity_rapid);
 
         // Set up Realm
-        realmConfig = new RealmConfiguration.Builder(this)
+        Realm.init(getApplicationContext());
+        realmConfig = new RealmConfiguration.Builder()
                 .schemaVersion(1)
                 .deleteRealmIfMigrationNeeded()
                 .build();
